Simplifies behavior of 'confirm'.
In Python>3, `True` and `False` are keywords and will always be
equal to 1 and 0, respectively. As a result, the previous implementation
of confirm would always pass the test `abort_on_unconfirmed in (True, 0)`
regardless of the value of `abort_on_unconfirmed`.

The more complex behavior triggered when `abort_on_unconfirmed` is of
type `int` or `str` does not appear to be in use and is removed to
correct this issue.

Refs: https://stackoverflow.com/questions/2764017/is-false-0-and-true-1-in-python-an-implementation-detail-or-is-it-guarante#2764099